+++ src/wamytmapp/tests/__init__.py
@@ -1,5 +1,7 @@
 from .addtimetangeform import *
+from .conflict import *
 from .oudelegate import *
 from .selectlistitems import *
 from .timerangemanager import *
+from .timerangemodel import *
 from .views import *